requirements.txt
git+https://github.com/AngellusMortis/python-valve
discord.py==1.5.1
